move tests clean up
