+++ tests/genomics_tests/test_genomic_job_controller.py
@@ -5,11 +5,10 @@ from rdr_service import clock, config
 from rdr_service.api_util import open_cloud_file
 from rdr_service.clock import FakeClock
 from rdr_service.dao.genomics_dao import GenomicGcDataFileDao, GenomicGCValidationMetricsDao, GenomicIncidentDao, \
-    GenomicInformingLoopDao, GenomicResultViewedDao, GenomicSetMemberDao, UserEventMetricsDao, GenomicJobRunDao, \
-    GenomicMemberReportStateDao
+    GenomicSetMemberDao, UserEventMetricsDao, GenomicJobRunDao
 from rdr_service.dao.message_broker_dao import MessageBrokenEventDataDao
 from rdr_service.genomic_enums import GenomicIncidentCode, GenomicJob, GenomicWorkflowState, GenomicSubProcessResult, \
-    GenomicSubProcessStatus, GenomicManifestTypes, GenomicQcStatus, GenomicReportState
+    GenomicSubProcessStatus, GenomicManifestTypes, GenomicQcStatus
 from rdr_service.genomic.genomic_job_components import GenomicFileIngester
 from rdr_service.genomic.genomic_job_controller import GenomicJobController
 from rdr_service.model.genomics import GenomicGcDataFile, GenomicIncident, GenomicSetMember, GenomicGCValidationMetrics
@@ -23,13 +22,10 @@ class GenomicJobControllerTest(BaseTestCase):
         self.data_file_dao = GenomicGcDataFileDao()
         self.event_data_dao = MessageBrokenEventDataDao()
         self.incident_dao = GenomicIncidentDao()
-        self.informing_loop_dao = GenomicInformingLoopDao()
-        self.result_viewed_dao = GenomicResultViewedDao()
         self.member_dao = GenomicSetMemberDao()
         self.metrics_dao = GenomicGCValidationMetricsDao()
         self.user_event_metrics_dao = UserEventMetricsDao()
         self.job_run_dao = GenomicJobRunDao()
-        self.report_state_dao = GenomicMemberReportStateDao()
 
     def test_incident_with_long_message(self):
         """Make sure the length of incident messages doesn't cause issues when recording them"""
@@ -252,517 +248,6 @@ class GenomicJobControllerTest(BaseTestCase):
             self.assertEqual(expected_objs[i].ignore_flag, inserted_files[i].ignore_flag)
             self.assertEqual(expected_objs[i].metadata, inserted_files[i].metadata)
             self.assertEqual(expected_objs[i].modified, inserted_files[i].modified)
-
-    def test_informing_loop_ingestion_message_broker(self):
-        loop_decision = 'informing_loop_decision'
-        loop_started = 'informing_loop_started'
-        # https://docs.google.com/document/d/1E1tNSi1mWwhBSCs9Syprbzl5E0SH3c_9oLduG1mzlcY/edit#heading=h.2m73apfm9irj
-        loop_module_types = ['gem', 'hdr', 'pgx']
-
-        gen_set = self.data_generator.create_database_genomic_set(
-            genomicSetName=".",
-            genomicSetCriteria=".",
-            genomicSetVersion=1
-        )
-
-        participant = self.data_generator.create_database_participant()
-        sample_id = '22222222'
-
-        self.data_generator.create_database_genomic_set_member(
-            genomicSetId=gen_set.id,
-            sampleId=sample_id,
-            participantId=participant.participantId,
-            biobankId="1",
-            genomeType="aou_array",
-            genomicWorkflowState=GenomicWorkflowState.AW0
-        )
-
-        # gem array
-        message_broker_record = self.data_generator.create_database_message_broker_record(
-            participantId=participant.participantId,
-            eventType=loop_decision,
-            eventAuthoredTime=clock.CLOCK.now(),
-            messageOrigin='example@example.com',
-            requestBody={'module_type': loop_module_types[0], 'decision_value': 'yes'},
-            requestTime=clock.CLOCK.now(),
-            responseError='',
-            responseCode='200',
-            responseTime=clock.CLOCK.now()
-        )
-
-        for key, value in message_broker_record.requestBody.items():
-            self.data_generator.create_database_message_broker_event_data(
-                participantId=message_broker_record.participantId,
-                messageRecordId=message_broker_record.id,
-                eventType=message_broker_record.eventType,
-                eventAuthoredTime=message_broker_record.eventAuthoredTime,
-                fieldName=key,
-                valueString=value
-            )
-
-        with GenomicJobController(GenomicJob.INGEST_INFORMING_LOOP) as controller:
-            controller.ingest_records_from_message_broker_data(
-                message_record_id=message_broker_record.id,
-                event_type=loop_decision
-            )
-
-        decision_genomic_record = self.informing_loop_dao.get(1)
-
-        self.assertIsNotNone(decision_genomic_record)
-        self.assertIsNotNone(decision_genomic_record.event_type)
-        self.assertIsNotNone(decision_genomic_record.module_type)
-        self.assertIsNotNone(decision_genomic_record.decision_value)
-
-        self.assertEqual(decision_genomic_record.message_record_id, message_broker_record.id)
-        self.assertEqual(decision_genomic_record.participant_id, message_broker_record.participantId)
-        self.assertEqual(decision_genomic_record.event_type, loop_decision)
-        self.assertTrue(decision_genomic_record.module_type == 'gem')
-        self.assertEqual(decision_genomic_record.decision_value, 'yes')
-        self.assertEqual(decision_genomic_record.sample_id, sample_id)
-
-        message_broker_record_two = self.data_generator.create_database_message_broker_record(
-            participantId=participant.participantId,
-            eventType=loop_started,
-            eventAuthoredTime=clock.CLOCK.now(),
-            messageOrigin='example@example.com',
-            requestBody={'module_type': loop_module_types[0]},
-            requestTime=clock.CLOCK.now(),
-            responseError='',
-            responseCode='200',
-            responseTime=clock.CLOCK.now()
-        )
-
-        for key, value in message_broker_record_two.requestBody.items():
-            self.data_generator.create_database_message_broker_event_data(
-                participantId=message_broker_record_two.participantId,
-                messageRecordId=message_broker_record_two.id,
-                eventType=message_broker_record_two.eventType,
-                eventAuthoredTime=message_broker_record_two.eventAuthoredTime,
-                fieldName=key,
-                valueString=value
-            )
-
-        with GenomicJobController(GenomicJob.INGEST_INFORMING_LOOP) as controller:
-            controller.ingest_records_from_message_broker_data(
-                message_record_id=message_broker_record_two.id,
-                event_type=loop_started
-            )
-
-        started_genomic_record = self.informing_loop_dao.get(2)
-
-        self.assertIsNotNone(started_genomic_record)
-        self.assertIsNotNone(started_genomic_record.event_type)
-        self.assertIsNotNone(started_genomic_record.module_type)
-        self.assertIsNone(started_genomic_record.decision_value)
-
-        self.assertEqual(started_genomic_record.message_record_id, message_broker_record_two.id)
-        self.assertEqual(started_genomic_record.participant_id, message_broker_record_two.participantId)
-        self.assertEqual(started_genomic_record.event_type, loop_started)
-        self.assertEqual(started_genomic_record.module_type, loop_module_types[0])
-        self.assertEqual(decision_genomic_record.sample_id, sample_id)
-
-        sample_id = '22333333'
-
-        self.data_generator.create_database_genomic_set_member(
-            genomicSetId=gen_set.id,
-            sampleId=sample_id,
-            participantId=participant.participantId,
-            biobankId="2",
-            genomeType="aou_wgs",
-            genomicWorkflowState=GenomicWorkflowState.AW0
-        )
-
-        # hdr wgs
-        message_broker_record_three = self.data_generator.create_database_message_broker_record(
-            participantId=participant.participantId,
-            eventType=loop_decision,
-            eventAuthoredTime=clock.CLOCK.now(),
-            messageOrigin='example@example.com',
-            requestBody={'module_type': loop_module_types[1], 'decision_value': 'yes'},
-            requestTime=clock.CLOCK.now(),
-            responseError='',
-            responseCode='200',
-            responseTime=clock.CLOCK.now()
-        )
-
-        for key, value in message_broker_record_three.requestBody.items():
-            self.data_generator.create_database_message_broker_event_data(
-                participantId=message_broker_record_three.participantId,
-                messageRecordId=message_broker_record_three.id,
-                eventType=message_broker_record_three.eventType,
-                eventAuthoredTime=message_broker_record_three.eventAuthoredTime,
-                fieldName=key,
-                valueString=value
-            )
-
-        with GenomicJobController(GenomicJob.INGEST_INFORMING_LOOP) as controller:
-            controller.ingest_records_from_message_broker_data(
-                message_record_id=message_broker_record_three.id,
-                event_type=loop_decision
-            )
-
-        decision_genomic_record = self.informing_loop_dao.get(3)
-
-        self.assertIsNotNone(decision_genomic_record)
-        self.assertIsNotNone(decision_genomic_record.event_type)
-        self.assertIsNotNone(decision_genomic_record.module_type)
-        self.assertIsNotNone(decision_genomic_record.decision_value)
-
-        self.assertEqual(decision_genomic_record.message_record_id, message_broker_record_three.id)
-        self.assertEqual(decision_genomic_record.participant_id, message_broker_record_three.participantId)
-        self.assertEqual(decision_genomic_record.event_type, loop_decision)
-        self.assertTrue(decision_genomic_record.module_type == 'hdr')
-        self.assertEqual(decision_genomic_record.decision_value, 'yes')
-        self.assertEqual(decision_genomic_record.sample_id, sample_id)
-
-    def test_result_viewed_ingestion_message_broker(self):
-        event_type = 'result_viewed'
-        participant = self.data_generator.create_database_participant()
-        # https://docs.google.com/document/d/1E1tNSi1mWwhBSCs9Syprbzl5E0SH3c_9oLduG1mzlcY/edit#heading=h.dtikttz25h22
-        result_module_types = ['gem', 'hdr_v1', 'pgx_v1']
-
-        gen_set = self.data_generator.create_database_genomic_set(
-            genomicSetName=".",
-            genomicSetCriteria=".",
-            genomicSetVersion=1
-        )
-
-        sample_id = '2222222'
-
-        self.data_generator.create_database_genomic_set_member(
-            genomicSetId=gen_set.id,
-            sampleId=sample_id,
-            participantId=participant.participantId,
-            biobankId="1",
-            genomeType="aou_array",
-            genomicWorkflowState=GenomicWorkflowState.AW0
-        )
-
-        message_broker_record = self.data_generator.create_database_message_broker_record(
-            participantId=participant.participantId,
-            eventType=event_type,
-            eventAuthoredTime=clock.CLOCK.now(),
-            messageOrigin='example@example.com',
-            requestBody={'result_type': result_module_types[0]},
-            requestTime=clock.CLOCK.now(),
-            responseError='',
-            responseCode='200',
-            responseTime=clock.CLOCK.now()
-        )
-
-        for key, value in message_broker_record.requestBody.items():
-            self.data_generator.create_database_message_broker_event_data(
-                participantId=message_broker_record.participantId,
-                messageRecordId=message_broker_record.id,
-                eventType=message_broker_record.eventType,
-                eventAuthoredTime=message_broker_record.eventAuthoredTime,
-                fieldName=key,
-                valueString=value
-            )
-
-        with GenomicJobController(GenomicJob.INGEST_RESULT_VIEWED) as controller:
-            controller.ingest_records_from_message_broker_data(
-                message_record_id=message_broker_record.id,
-                event_type=event_type
-            )
-
-        result_viewed_genomic_record = self.result_viewed_dao.get_all()
-
-        self.assertIsNotNone(result_viewed_genomic_record)
-        self.assertEqual(len(result_viewed_genomic_record), 1)
-
-        result_viewed_genomic_record = result_viewed_genomic_record[0]
-
-        self.assertIsNotNone(result_viewed_genomic_record.event_type)
-        self.assertIsNotNone(result_viewed_genomic_record.module_type)
-
-        self.assertEqual(result_viewed_genomic_record.message_record_id, message_broker_record.id)
-        self.assertEqual(result_viewed_genomic_record.participant_id, message_broker_record.participantId)
-        self.assertEqual(result_viewed_genomic_record.event_type, event_type)
-        self.assertTrue(result_viewed_genomic_record.module_type == 'gem')
-        self.assertEqual(result_viewed_genomic_record.sample_id, sample_id)
-
-        self.assertEqual(result_viewed_genomic_record.first_viewed, message_broker_record.eventAuthoredTime)
-        self.assertEqual(result_viewed_genomic_record.last_viewed,  message_broker_record.eventAuthoredTime)
-
-        message_broker_record_two = self.data_generator.create_database_message_broker_record(
-            participantId=participant.participantId,
-            eventType=event_type,
-            eventAuthoredTime=clock.CLOCK.now() + datetime.timedelta(days=1),
-            messageOrigin='example@example.com',
-            requestBody={'result_type': result_viewed_genomic_record.module_type},
-            requestTime=clock.CLOCK.now(),
-            responseError='',
-            responseCode='200',
-            responseTime=clock.CLOCK.now()
-        )
-
-        for key, value in message_broker_record_two.requestBody.items():
-            self.data_generator.create_database_message_broker_event_data(
-                participantId=message_broker_record_two.participantId,
-                messageRecordId=message_broker_record_two.id,
-                eventType=message_broker_record_two.eventType,
-                eventAuthoredTime=message_broker_record_two.eventAuthoredTime,
-                fieldName=key,
-                valueString=value
-            )
-
-        with GenomicJobController(GenomicJob.INGEST_RESULT_VIEWED) as controller:
-            controller.ingest_records_from_message_broker_data(
-                message_record_id=message_broker_record_two.id,
-                event_type=event_type
-            )
-
-        result_viewed_genomic_record = self.result_viewed_dao.get_all()
-
-        self.assertIsNotNone(result_viewed_genomic_record)
-        self.assertEqual(len(result_viewed_genomic_record), 1)
-
-        result_viewed_genomic_record = result_viewed_genomic_record[0]
-
-        self.assertEqual(result_viewed_genomic_record.first_viewed, message_broker_record.eventAuthoredTime)
-
-        # check updated record has the last viewed time
-        self.assertEqual(result_viewed_genomic_record.last_viewed, message_broker_record_two.eventAuthoredTime)
-        self.assertEqual(result_viewed_genomic_record.message_record_id, message_broker_record.id)
-
-        sample_id = '223333'
-
-        self.data_generator.create_database_genomic_set_member(
-            genomicSetId=gen_set.id,
-            sampleId=sample_id,
-            participantId=participant.participantId,
-            biobankId="2",
-            genomeType="aou_wgs",
-            genomicWorkflowState=GenomicWorkflowState.AW0
-        )
-
-        message_broker_record_two = self.data_generator.create_database_message_broker_record(
-            participantId=participant.participantId,
-            eventType=event_type,
-            eventAuthoredTime=clock.CLOCK.now(),
-            messageOrigin='example@example.com',
-            requestBody={'result_type': result_module_types[1]},
-            requestTime=clock.CLOCK.now(),
-            responseError='',
-            responseCode='200',
-            responseTime=clock.CLOCK.now()
-        )
-
-        for key, value in message_broker_record_two.requestBody.items():
-            self.data_generator.create_database_message_broker_event_data(
-                participantId=message_broker_record_two.participantId,
-                messageRecordId=message_broker_record_two.id,
-                eventType=message_broker_record_two.eventType,
-                eventAuthoredTime=message_broker_record_two.eventAuthoredTime,
-                fieldName=key,
-                valueString=value
-            )
-
-        with GenomicJobController(GenomicJob.INGEST_RESULT_VIEWED) as controller:
-            controller.ingest_records_from_message_broker_data(
-                message_record_id=message_broker_record_two.id,
-                event_type=event_type
-            )
-
-        result_viewed_genomic_record = self.result_viewed_dao.get(2)
-
-        self.assertIsNotNone(result_viewed_genomic_record)
-
-        self.assertIsNotNone(result_viewed_genomic_record.event_type)
-        self.assertIsNotNone(result_viewed_genomic_record.module_type)
-
-        self.assertEqual(result_viewed_genomic_record.message_record_id, message_broker_record_two.id)
-        self.assertEqual(result_viewed_genomic_record.participant_id, message_broker_record_two.participantId)
-        self.assertEqual(result_viewed_genomic_record.event_type, event_type)
-        self.assertTrue(result_viewed_genomic_record.module_type == 'hdr_v1')
-        self.assertEqual(result_viewed_genomic_record.sample_id, sample_id)
-
-        self.assertEqual(result_viewed_genomic_record.first_viewed, message_broker_record_two.eventAuthoredTime)
-        self.assertEqual(result_viewed_genomic_record.last_viewed, message_broker_record_two.eventAuthoredTime)
-
-    def test_result_ready_ingestion_message_broker(self):
-        event_type = 'result_ready'
-        participant = self.data_generator.create_database_participant()
-        # https://docs.google.com/document/d/1E1tNSi1mWwhBSCs9Syprbzl5E0SH3c_9oLduG1mzlcY/edit#heading=h.dtikttz25h22
-        result_module_types = ['hdr_v1', 'pgx_v1']
-
-        gen_set = self.data_generator.create_database_genomic_set(
-            genomicSetName=".",
-            genomicSetCriteria=".",
-            genomicSetVersion=1
-        )
-
-        sample_id = '2222222'
-
-        self.data_generator.create_database_genomic_set_member(
-            genomicSetId=gen_set.id,
-            sampleId=sample_id,
-            participantId=participant.participantId,
-            biobankId="1",
-            genomeType="aou_wgs",
-            genomicWorkflowState=GenomicWorkflowState.AW0
-        )
-
-        # HDR Positive Records
-        message_broker_record_hdr_positive = self.data_generator.create_database_message_broker_record(
-            participantId=participant.participantId,
-            eventType=event_type,
-            eventAuthoredTime=clock.CLOCK.now(),
-            messageOrigin='example@example.com',
-            requestBody={
-                'result_type': result_module_types[0],
-                'hdr_result_status': 'positive',
-                'report_revision_number': 0
-            },
-            requestTime=clock.CLOCK.now(),
-            responseError='',
-            responseCode='200',
-            responseTime=clock.CLOCK.now()
-        )
-
-        for key, value in message_broker_record_hdr_positive.requestBody.items():
-            self.data_generator.create_database_message_broker_event_data(
-                participantId=message_broker_record_hdr_positive.participantId,
-                messageRecordId=message_broker_record_hdr_positive.id,
-                eventType=message_broker_record_hdr_positive.eventType,
-                eventAuthoredTime=message_broker_record_hdr_positive.eventAuthoredTime,
-                fieldName=key,
-                valueString=value
-            )
-
-        with GenomicJobController(GenomicJob.INGEST_RESULT_READY) as controller:
-            controller.ingest_records_from_message_broker_data(
-                message_record_id=message_broker_record_hdr_positive.id,
-                event_type=event_type
-            )
-
-        result_ready_genomic_record = self.report_state_dao.get_all()
-
-        self.assertIsNotNone(result_ready_genomic_record)
-        self.assertEqual(len(result_ready_genomic_record), 1)
-
-        result_ready_genomic_record = result_ready_genomic_record[0]
-
-        self.assertIsNotNone(result_ready_genomic_record.event_type)
-        self.assertIsNotNone(result_ready_genomic_record.module)
-
-        self.assertEqual(result_ready_genomic_record.message_record_id, message_broker_record_hdr_positive.id)
-        self.assertEqual(result_ready_genomic_record.participant_id, message_broker_record_hdr_positive.participantId)
-        self.assertEqual(result_ready_genomic_record.event_type, event_type)
-        self.assertTrue(result_ready_genomic_record.module == result_module_types[0])
-        self.assertEqual(result_ready_genomic_record.sample_id, sample_id)
-
-        # check for correct report state
-        self.assertEqual(result_ready_genomic_record.genomic_report_state, GenomicReportState.HDR_RPT_POSITIVE)
-        self.assertEqual(result_ready_genomic_record.genomic_report_state_str, GenomicReportState.HDR_RPT_POSITIVE.name)
-
-        # HDR Uninformative Records
-        message_broker_record_hdr_uninformative = self.data_generator.create_database_message_broker_record(
-            participantId=participant.participantId,
-            eventType=event_type,
-            eventAuthoredTime=clock.CLOCK.now(),
-            messageOrigin='example@example.com',
-            requestBody={
-                'result_type': result_module_types[0],
-                'hdr_result_status': 'uninformative',
-                'report_revision_number': 0
-            },
-            requestTime=clock.CLOCK.now(),
-            responseError='',
-            responseCode='200',
-            responseTime=clock.CLOCK.now()
-        )
-
-        for key, value in message_broker_record_hdr_uninformative.requestBody.items():
-            self.data_generator.create_database_message_broker_event_data(
-                participantId=message_broker_record_hdr_uninformative.participantId,
-                messageRecordId=message_broker_record_hdr_uninformative.id,
-                eventType=message_broker_record_hdr_uninformative.eventType,
-                eventAuthoredTime=message_broker_record_hdr_uninformative.eventAuthoredTime,
-                fieldName=key,
-                valueString=value
-            )
-
-        with GenomicJobController(GenomicJob.INGEST_RESULT_READY) as controller:
-            controller.ingest_records_from_message_broker_data(
-                message_record_id=message_broker_record_hdr_uninformative.id,
-                event_type=event_type
-            )
-
-        result_ready_genomic_record = self.report_state_dao.get_all()
-
-        self.assertIsNotNone(result_ready_genomic_record)
-        self.assertEqual(len(result_ready_genomic_record), 2)
-
-        result_ready_genomic_record = result_ready_genomic_record[1]
-
-        self.assertIsNotNone(result_ready_genomic_record.event_type)
-        self.assertIsNotNone(result_ready_genomic_record.module)
-
-        self.assertEqual(result_ready_genomic_record.message_record_id, message_broker_record_hdr_uninformative.id)
-        self.assertEqual(result_ready_genomic_record.participant_id,
-                         message_broker_record_hdr_uninformative.participantId)
-        self.assertEqual(result_ready_genomic_record.event_type, event_type)
-        self.assertTrue(result_ready_genomic_record.module == result_module_types[0])
-        self.assertEqual(result_ready_genomic_record.sample_id, sample_id)
-
-        # check for correct report state
-        self.assertEqual(result_ready_genomic_record.genomic_report_state, GenomicReportState.HDR_RPT_UNINFORMATIVE)
-        self.assertEqual(result_ready_genomic_record.genomic_report_state_str,
-                         GenomicReportState.HDR_RPT_UNINFORMATIVE.name)
-
-        # PGX Records
-        message_broker_record_pgx = self.data_generator.create_database_message_broker_record(
-            participantId=participant.participantId,
-            eventType=event_type,
-            eventAuthoredTime=clock.CLOCK.now(),
-            messageOrigin='example@example.com',
-            requestBody={
-                'result_type': result_module_types[1],
-                'report_revision_number': 0
-            },
-            requestTime=clock.CLOCK.now(),
-            responseError='',
-            responseCode='200',
-            responseTime=clock.CLOCK.now()
-        )
-
-        for key, value in message_broker_record_pgx.requestBody.items():
-            self.data_generator.create_database_message_broker_event_data(
-                participantId=message_broker_record_pgx.participantId,
-                messageRecordId=message_broker_record_pgx.id,
-                eventType=message_broker_record_pgx.eventType,
-                eventAuthoredTime=message_broker_record_pgx.eventAuthoredTime,
-                fieldName=key,
-                valueString=value
-            )
-
-        with GenomicJobController(GenomicJob.INGEST_RESULT_READY) as controller:
-            controller.ingest_records_from_message_broker_data(
-                message_record_id=message_broker_record_pgx.id,
-                event_type=event_type
-            )
-
-        result_ready_genomic_record = self.report_state_dao.get_all()
-
-        self.assertIsNotNone(result_ready_genomic_record)
-        self.assertEqual(len(result_ready_genomic_record), 3)
-
-        result_ready_genomic_record = result_ready_genomic_record[2]
-
-        self.assertIsNotNone(result_ready_genomic_record.event_type)
-        self.assertIsNotNone(result_ready_genomic_record.module)
-
-        self.assertEqual(result_ready_genomic_record.message_record_id, message_broker_record_pgx.id)
-        self.assertEqual(result_ready_genomic_record.participant_id, message_broker_record_pgx.participantId)
-        self.assertEqual(result_ready_genomic_record.event_type, event_type)
-        self.assertTrue(result_ready_genomic_record.module == result_module_types[1])
-        self.assertEqual(result_ready_genomic_record.sample_id, sample_id)
-
-        # check for correct report state
-        self.assertEqual(result_ready_genomic_record.genomic_report_state, GenomicReportState.PGX_RPT_READY)
-        self.assertEqual(result_ready_genomic_record.genomic_report_state_str, GenomicReportState.PGX_RPT_READY.name)
 
     def test_updating_members_blocklists(self):
